rename viewholder package to controllers, ActivityPlayerPanel to ActivityPlayerController, PlayerPanel to PlayerController
